Change h2 headings from inline-block to block
Delete `display: inline-block` from CSS for h2 elements changing
`display` to default `block`. Add `width: fit-content`.

`display: inline-block` was used to prevent the background color from
stretching the whole width of the container. `display: block` with
`width: fit-content` is a simpler and more elegant way to achieve the
desired outcome. `display: inline-block` only works if the heading is
not next to an inline or inline-block element.